backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
pytest
httpx
pytest-asyncio
pydantic-settings
requests